requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
albumentations>=1.3.1
opencv-python-headless>=4.8.0
scikit-learn>=1.3.2
pandas>=2.1.0
numpy>=1.24.0
tqdm>=4.66.1
Pillow>=10.0.0
